requirements.txt
appdirs==1.4.3
aspy.yaml==1.3.0
astroid==2.3.3
attrs==19.3.0
black==19.10b0
cfgv==2.0.1
Click==7.0
Django==2.2.7
djangorestframework==3.10.3
identify==1.4.7
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3
pathspec==0.6.0
pre-commit==1.20.0
psycopg2-binary==2.8.4
pylint==2.4.4
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
virtualenv==16.7.7
wrapt==1.11.2
zipp==0.6.0
